fix grapple explosion bug on doors
Avoid exploding the hook because we cannot launch it anymore after that.
It's not easy to find how to reproduce it:
you can do it near the start of the third level on the moving pylons,
when the pylon just starts to go out of the wall, launch a grapple on its side.